# src/public/swagger/Auth.yaml
openapi: 3.0.0
info:
  title: School Management API
  version: 1.0.0
  description: API documentation for the School Management System, including authentication and core entities like Users, Kelas, Siswa, and Guru.
...
paths:
...
  /api/v1/users:
    get:
      summary: Get all users
      tags: [Users]
      responses:
        200:
          description: A list of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListUsersResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    post:
      summary: Create a new user (for admin use, different from register)
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        201:
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/mata-pelajaran/{id_mapel}:
    get:
      summary: Get a mata-pelajaran by ID
      tags: [Mata Pelajaran]
      parameters:
        - in: path
          name: id_mapel
          schema:
            type: integer
          required: true
          description: Numeric ID of the mata-pelajaran to retrieve
      responses:
        200:
          description: Subject found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MataPelajaranResponse'
        404:
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    put:
      summary: Update a mata-pelajaran by ID
      tags: [Mata Pelajaran]
      parameters:
        - in: path
          name: id_mapel
          schema:
            type: integer
          required: true
          description: Numeric ID of the mata-pelajaran to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MataPelajaran'
      responses:
        200:
          description: Subject updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MataPelajaranResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        404:
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    delete:
      summary: Delete a mata-pelajaran by ID
      tags: [Mata Pelajaran]
      parameters:
        - in: path
          name: id_mapel
          schema:
            type: integer
          required: true
          description: Numeric ID of the mata-pelajaran to delete
      responses:
        200:
          description: Subject deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        404:
          description: Subject not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
  /api/v1/siswa/{id_siswa}:
    get:
      summary: Get a siswa by ID
      tags: [Siswa]
      parameters:
        - in: path
          name: id_siswa
          schema:
            type: integer
          required: true
          description: Numeric ID of the siswa to retrieve
      responses:
        200:
          description: Student found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiswaResponse'
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    put:
      summary: Update a siswa by ID
      tags: [Siswa]
      parameters:
        - in: path
          name: id_siswa
          schema:
            type: integer
          required: true
          description: Numeric ID of the siswa to update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Siswa'
      responses:
        200:
          description: Student updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiswaResponse'
        400:
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
    delete:
      summary: Delete a siswa by ID
      tags: [Siswa]
      parameters:
        - in: path
          name: id_siswa
          schema:
            type: integer
          required: true
          description: Numeric ID of the siswa to delete
      responses:
        200:
          description: Student deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardSuccessResponse'
        404:
          description: Student not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        401:
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
        500:
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StandardErrorResponse'
      security:
        - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StandardSuccessResponse:
      type: object
      properties:
        status:
          type: boolean
          description: Indicates whether the request was successful
          example: true
        message:
          type: string
          description: A descriptive message
          example: Operation successful
      required:
        - status
        - message
    StandardErrorResponse:
      type: object
      properties:
        status:
          type: boolean
          description: Indicates whether the request was successful
          example: false
        message:
          type: string
          description: A descriptive error message
          example: An error occurred
      required:
        - status
        - message

    # --- Core Entity Schemas (based on Prisma) ---
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
          description: Unique identifier for the user.
          example: 1
        username:
          type: string
          description: Unique username for the user.
          example: "siswa001"
        name:
          type: string
          description: Full name of the user.
          example: "Budi Santoso"
        role:
          type: string
          enum: [admin, guru, siswa]
          description: Role of the user.
          example: siswa
        createdAt:
          type: string
          format: date-time
          readOnly: true
          description: Date and time when the user was created.
          example: "2024-01-01T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          description: Date and time when the user was last updated.
          example: "2024-01-01T10:00:00Z"
    CreateUser:
      type: object
      required:
        - username
        - name
        - role
        - password
      properties:
        username:
          type: string
          description: Unique username for the user.
          example: "admin01"
        name:
          type: string
          description: Full name of the user.
          example: "Admin System"
        role:
          type: string
          enum: [admin, guru, siswa]
          description: Role of the user.
          example: admin
        password:
          type: string
          format: password
          minLength: 8
          description: Password for the user account.
          example: "strongpassword"
...
    UserResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: User fetched successfully
        data:
          $ref: '#/components/schemas/User'
    ListUsersResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Users fetched successfully
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    MataPelajaran:
      type: object
      required:
        - nama_mapel
      properties:
        id_mapel:
          type: integer
          readOnly: true
          description: Unique identifier for the mata-pelajaran.
          example: 1
        nama_mapel:
          type: string
          description: Name of the mata-pelajaran (e.g., "Matematika", "Fisika", "Bahasa Indonesia").
          example: "Matematika"
      example:
        nama_mapel: "Fisika"
    MataPelajaranResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Subject fetched successfully
        data:
          $ref: '#/components/schemas/MataPelajaran'
...
    Siswa:
      type: object
      required:
        - nisn
        - nama_siswa
        - no_absen
        - id_kelas
        - id_jurusan
      properties:
        id_siswa:
          type: integer
          readOnly: true
          description: Unique identifier for the siswa.
          example: 1
        nisn:
          type: string
          description: National Student Identification Number (Nomor Induk Siswa Nasional) for the siswa.
          example: "0012345678"
        nama_siswa:
          type: string
          description: Full name of the siswa.
          example: "Rudi Haryanto"
        no_absen:
          type: integer
          description: Absen number of the siswa in their kelas.
          example: 5
        id_kelas:
          type: integer
          description: ID of the kelas the siswa belongs to.
          example: 1
        id_jurusan:
          type: integer
          description: ID of the jurusan the siswa belongs to.
          example: 2
      example:
        nisn: "0012345678"
        nama_siswa: "Rudi Haryanto"
        no_absen: 5
        id_kelas: 1
        id_jurusan: 2
    SiswaResponse:
      type: object
      properties:
        status:
          type: boolean
          example: true
        message:
          type: string
          example: Student fetched successfully
        data:
          $ref: '#/components/schemas/Siswa'
...